+++ backend/formation/src/main/java/com/s3m/formation/api/kpi/client/projection/TotalFormationHoursProjection.java
@@ -1,5 +1,7 @@
 package com.s3m.formation.api.kpi.client.projection;
 
+import java.math.BigDecimal;
+
 public interface TotalFormationHoursProjection {
-    Double getTotalHeures();
+    BigDecimal getTotalHeures();
 }
